Joystick Functions Class
The first program I've decided to code of 100DaysofCode. This programs works on the Raspberry Pi with a Sense HAT. It takes in input from the joystick and navigates an 8 x 8 matrix. Ideally for the LED matrix on the Sense HAT. It's not finished quite yet, still need to make an LED move on the matrix. I plan on using this class in other programs through this challenge.